requirements.txt
python-docx 1.1.2
stanza 1.10.1